% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{accounts.mfaEnrollment.start}
\alias{accounts.mfaEnrollment.start}
\title{Step one of the MFA enrollment process. In SMS case, this sends an SMS verification code to the user.}
\usage{
accounts.mfaEnrollment.start(
  GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest}{The \link{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest functions: 
\code{\link{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest}()}
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest functions}
