% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{accounts.mfaEnrollment.withdraw}
\alias{accounts.mfaEnrollment.withdraw}
\title{Revokes one second factor from the enrolled second factors for an account.}
\usage{
accounts.mfaEnrollment.withdraw(GoogleCloudIdentitytoolkitV2WithdrawMfaRequest)
}
\arguments{
\item{GoogleCloudIdentitytoolkitV2WithdrawMfaRequest}{The \link{GoogleCloudIdentitytoolkitV2WithdrawMfaRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitV2WithdrawMfaRequest functions: 
\code{\link{GoogleCloudIdentitytoolkitV2WithdrawMfaRequest}()}
}
\concept{GoogleCloudIdentitytoolkitV2WithdrawMfaRequest functions}
