% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{accounts.revokeToken}
\alias{accounts.revokeToken}
\title{Revokes a user's token from an Identity Provider (IdP). This is done by manually providing an IdP credential, and the token types for revocation. An \href{https://cloud.google.com/docs/authentication/api-keys}{API key} is required in the request in order to identify the Google Cloud project.}
\usage{
accounts.revokeToken(GoogleCloudIdentitytoolkitV2RevokeTokenRequest)
}
\arguments{
\item{GoogleCloudIdentitytoolkitV2RevokeTokenRequest}{The \link{GoogleCloudIdentitytoolkitV2RevokeTokenRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitV2RevokeTokenRequest functions: 
\code{\link{GoogleCloudIdentitytoolkitV2RevokeTokenRequest}()}
}
\concept{GoogleCloudIdentitytoolkitV2RevokeTokenRequest functions}
