% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{getRecaptchaConfig}
\alias{getRecaptchaConfig}
\title{Gets parameters needed for reCAPTCHA analysis.}
\usage{
getRecaptchaConfig(version = NULL, clientType = NULL, tenantId = NULL)
}
\arguments{
\item{version}{The reCAPTCHA version}

\item{clientType}{reCAPTCHA Enterprise uses separate site keys for different client types}

\item{tenantId}{The id of a tenant}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}
}
