% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2AllowByDefault}
\alias{GoogleCloudIdentitytoolkitAdminV2AllowByDefault}
\title{GoogleCloudIdentitytoolkitAdminV2AllowByDefault Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2AllowByDefault(disallowedRegions = NULL)
}
\arguments{
\item{disallowedRegions}{Two letter unicode region codes to disallow as defined by https://cldr}
}
\value{
GoogleCloudIdentitytoolkitAdminV2AllowByDefault object
}
\description{
GoogleCloudIdentitytoolkitAdminV2AllowByDefault Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a policy of allowing every region by default and adding disallowed regions to a disallow list.
}
\concept{GoogleCloudIdentitytoolkitAdminV2AllowByDefault functions}
