% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2AllowlistOnly}
\alias{GoogleCloudIdentitytoolkitAdminV2AllowlistOnly}
\title{GoogleCloudIdentitytoolkitAdminV2AllowlistOnly Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2AllowlistOnly(allowedRegions = NULL)
}
\arguments{
\item{allowedRegions}{Two letter unicode region codes to allow as defined by https://cldr}
}
\value{
GoogleCloudIdentitytoolkitAdminV2AllowlistOnly object
}
\description{
GoogleCloudIdentitytoolkitAdminV2AllowlistOnly Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a policy of only allowing regions by explicitly adding them to an allowlist.
}
\concept{GoogleCloudIdentitytoolkitAdminV2AllowlistOnly functions}
