% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Anonymous}
\alias{GoogleCloudIdentitytoolkitAdminV2Anonymous}
\title{GoogleCloudIdentitytoolkitAdminV2Anonymous Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Anonymous(enabled = NULL)
}
\arguments{
\item{enabled}{Whether anonymous user auth is enabled for the project or not}
}
\value{
GoogleCloudIdentitytoolkitAdminV2Anonymous object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Anonymous Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration options related to authenticating an anonymous user.
}
\concept{GoogleCloudIdentitytoolkitAdminV2Anonymous functions}
