% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2AppleSignInConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2AppleSignInConfig}
\title{GoogleCloudIdentitytoolkitAdminV2AppleSignInConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2AppleSignInConfig(
  codeFlowConfig = NULL,
  bundleIds = NULL
)
}
\arguments{
\item{codeFlowConfig}{No description}

\item{bundleIds}{A list of Bundle ID's usable by this project}
}
\value{
GoogleCloudIdentitytoolkitAdminV2AppleSignInConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2AppleSignInConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional config for SignInWithApple.
}
\concept{GoogleCloudIdentitytoolkitAdminV2AppleSignInConfig functions}
