% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig}
\title{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig(
  GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers = NULL,
  triggers = NULL,
  forwardInboundCredentials = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers}{The \link{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers} object or list of objects}

\item{triggers}{Map of Trigger to event type}

\item{forwardInboundCredentials}{The user credentials to include in the JWT payload that is sent to the registered Blocking Functions}
}
\value{
GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration related to Blocking Functions.
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig functions}
