% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers}
\alias{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers}
\title{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers()
}
\value{
GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers object
}
\description{
GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig.triggers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Map of Trigger to event type. Key should be one of the supported event types: 'beforeCreate', 'beforeSignIn'
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2BlockingFunctionsConfig functions}
