% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ClientConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2ClientConfig}
\title{GoogleCloudIdentitytoolkitAdminV2ClientConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ClientConfig(permissions = NULL)
}
\arguments{
\item{permissions}{Configuration related to restricting a user's ability to affect their account}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ClientConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ClientConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options related to how clients making requests on behalf of a project should be configured.
}
\concept{GoogleCloudIdentitytoolkitAdminV2ClientConfig functions}
