% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ClientPermissionConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2ClientPermissionConfig}
\title{GoogleCloudIdentitytoolkitAdminV2ClientPermissionConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ClientPermissionConfig(permissions = NULL)
}
\arguments{
\item{permissions}{Configuration related to restricting a user's ability to affect their account}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ClientPermissionConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ClientPermissionConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options related to how clients making requests on behalf of a tenant should be configured.
}
\concept{GoogleCloudIdentitytoolkitAdminV2ClientPermissionConfig functions}
