% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ClientPermissions}
\alias{GoogleCloudIdentitytoolkitAdminV2ClientPermissions}
\title{GoogleCloudIdentitytoolkitAdminV2ClientPermissions Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ClientPermissions(
  disabledUserSignup = NULL,
  disabledUserDeletion = NULL
)
}
\arguments{
\item{disabledUserSignup}{When true, end users cannot sign up for a new account on the associated project through any of our API methods}

\item{disabledUserDeletion}{When true, end users cannot delete their account on the associated project through any of our API methods}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ClientPermissions object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ClientPermissions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration related to restricting a user's ability to affect their account.
}
\concept{GoogleCloudIdentitytoolkitAdminV2ClientPermissions functions}
