% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2CodeFlowConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2CodeFlowConfig}
\title{GoogleCloudIdentitytoolkitAdminV2CodeFlowConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2CodeFlowConfig(
  teamId = NULL,
  privateKey = NULL,
  keyId = NULL
)
}
\arguments{
\item{teamId}{Apple Developer Team ID}

\item{privateKey}{Private key used for signing the client secret JWT}

\item{keyId}{Key ID for the private key}
}
\value{
GoogleCloudIdentitytoolkitAdminV2CodeFlowConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2CodeFlowConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional config for Apple for code flow.
}
\concept{GoogleCloudIdentitytoolkitAdminV2CodeFlowConfig functions}
