% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Config}
\alias{GoogleCloudIdentitytoolkitAdminV2Config}
\title{GoogleCloudIdentitytoolkitAdminV2Config Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Config(
  notification = NULL,
  signIn = NULL,
  authorizedDomains = NULL,
  smsRegionConfig = NULL,
  mfa = NULL,
  autodeleteAnonymousUsers = NULL,
  monitoring = NULL,
  passwordPolicyConfig = NULL,
  recaptchaConfig = NULL,
  quota = NULL,
  blockingFunctions = NULL,
  multiTenant = NULL,
  client = NULL,
  emailPrivacyConfig = NULL
)
}
\arguments{
\item{notification}{Configuration related to sending notifications to users}

\item{signIn}{Configuration related to local sign in methods}

\item{authorizedDomains}{List of domains authorized for OAuth redirects}

\item{smsRegionConfig}{Configures which regions are enabled for SMS verification code sending}

\item{mfa}{Configuration for this project's multi-factor authentication, including whether it is active and what factors can be used for the second factor}

\item{autodeleteAnonymousUsers}{Whether anonymous users will be auto-deleted after a period of 30 days}

\item{monitoring}{Configuration related to monitoring project activity}

\item{passwordPolicyConfig}{The project level password policy configuration}

\item{recaptchaConfig}{The project-level reCAPTCHA config}

\item{quota}{Configuration related to quotas}

\item{blockingFunctions}{Configuration related to blocking functions}

\item{multiTenant}{Configuration related to multi-tenant functionality}

\item{client}{Options related to how clients making requests on behalf of a project should be configured}

\item{emailPrivacyConfig}{Configuration for settings related to email privacy and public visibility}
}
\value{
GoogleCloudIdentitytoolkitAdminV2Config object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Config Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an Identity Toolkit project.
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2Config functions: 
\code{\link{projects.updateConfig}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2Config functions}
