% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2CustomStrengthOptions}
\alias{GoogleCloudIdentitytoolkitAdminV2CustomStrengthOptions}
\title{GoogleCloudIdentitytoolkitAdminV2CustomStrengthOptions Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2CustomStrengthOptions(
  minPasswordLength = NULL,
  containsUppercaseCharacter = NULL,
  containsNumericCharacter = NULL,
  containsNonAlphanumericCharacter = NULL,
  containsLowercaseCharacter = NULL,
  maxPasswordLength = NULL
)
}
\arguments{
\item{minPasswordLength}{Minimum password length}

\item{containsUppercaseCharacter}{The password must contain an upper case character}

\item{containsNumericCharacter}{The password must contain a number}

\item{containsNonAlphanumericCharacter}{The password must contain a non alpha numeric character}

\item{containsLowercaseCharacter}{The password must contain a lower case character}

\item{maxPasswordLength}{Maximum password length}
}
\value{
GoogleCloudIdentitytoolkitAdminV2CustomStrengthOptions object
}
\description{
GoogleCloudIdentitytoolkitAdminV2CustomStrengthOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Custom strength options to enforce on user passwords.
}
\concept{GoogleCloudIdentitytoolkitAdminV2CustomStrengthOptions functions}
