% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdp}
\alias{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdp}
\title{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdp Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdp(
  idpId = NULL,
  description = NULL
)
}
\arguments{
\item{idpId}{Id the of Idp}

\item{description}{Description of the Idp}
}
\value{
GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdp object
}
\description{
GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdp Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Standard Identity Toolkit-trusted IDPs.
}
\concept{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdp functions}
