% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig}
\title{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig(
  enabled = NULL,
  appleSignInConfig = NULL,
  clientSecret = NULL,
  clientId = NULL,
  name = NULL
)
}
\arguments{
\item{enabled}{True if allows the user to sign in with the provider}

\item{appleSignInConfig}{Additional config for Apple-based projects}

\item{clientSecret}{OAuth client secret}

\item{clientId}{OAuth client ID}

\item{name}{The name of the DefaultSupportedIdpConfig resource, for example: 'projects/my-awesome-project/defaultSupportedIdpConfigs/google}
}
\value{
GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configurations options for authenticating with a the standard set of Identity Toolkit-trusted IDPs.
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig functions: 
\code{\link{projects.defaultSupportedIdpConfigs.create}()},
\code{\link{projects.defaultSupportedIdpConfigs.patch}()},
\code{\link{projects.tenants.defaultSupportedIdpConfigs.create}()},
\code{\link{projects.tenants.defaultSupportedIdpConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig functions}
