% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2DnsInfo}
\alias{GoogleCloudIdentitytoolkitAdminV2DnsInfo}
\title{GoogleCloudIdentitytoolkitAdminV2DnsInfo Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2DnsInfo(useCustomDomain = NULL)
}
\arguments{
\item{useCustomDomain}{Whether to use custom domain}
}
\value{
GoogleCloudIdentitytoolkitAdminV2DnsInfo object
}
\description{
GoogleCloudIdentitytoolkitAdminV2DnsInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information of custom domain DNS verification. By default, default_domain will be used. A custom domain can be configured using VerifyCustomDomain.
}
\concept{GoogleCloudIdentitytoolkitAdminV2DnsInfo functions}
