% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Email}
\alias{GoogleCloudIdentitytoolkitAdminV2Email}
\title{GoogleCloudIdentitytoolkitAdminV2Email Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Email(passwordRequired = NULL, enabled = NULL)
}
\arguments{
\item{passwordRequired}{Whether a password is required for email auth or not}

\item{enabled}{Whether email auth is enabled for the project or not}
}
\value{
GoogleCloudIdentitytoolkitAdminV2Email object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Email Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration options related to authenticating a user by their email address.
}
\concept{GoogleCloudIdentitytoolkitAdminV2Email functions}
