% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2EmailPrivacyConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2EmailPrivacyConfig}
\title{GoogleCloudIdentitytoolkitAdminV2EmailPrivacyConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2EmailPrivacyConfig(
  enableImprovedEmailPrivacy = NULL
)
}
\arguments{
\item{enableImprovedEmailPrivacy}{Migrates the project to a state of improved email privacy}
}
\value{
GoogleCloudIdentitytoolkitAdminV2EmailPrivacyConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2EmailPrivacyConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for settings related to email privacy and public visibility. Settings in this config protect against email enumeration, but may make some trade-offs in user-friendliness.
}
\concept{GoogleCloudIdentitytoolkitAdminV2EmailPrivacyConfig functions}
