% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2EmailTemplate}
\alias{GoogleCloudIdentitytoolkitAdminV2EmailTemplate}
\title{GoogleCloudIdentitytoolkitAdminV2EmailTemplate Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2EmailTemplate(
  replyTo = NULL,
  senderLocalPart = NULL,
  body = NULL,
  subject = NULL,
  bodyFormat = NULL,
  senderDisplayName = NULL
)
}
\arguments{
\item{replyTo}{Reply-to address}

\item{senderLocalPart}{Local part of From address}

\item{body}{Email body}

\item{subject}{Subject of the email}

\item{bodyFormat}{Email body format}

\item{senderDisplayName}{Sender display name}
}
\value{
GoogleCloudIdentitytoolkitAdminV2EmailTemplate object
}
\description{
GoogleCloudIdentitytoolkitAdminV2EmailTemplate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Email template. The subject and body fields can contain the following placeholders which will be replaced with the appropriate values: \%LINK\% - The link to use to redeem the send OOB code. \%EMAIL\% - The email where the email is being sent. \%NEW_EMAIL\% - The new email being set for the account (when applicable). \%APP_NAME\% - The GCP project's display name. \%DISPLAY_NAME\% - The user's display name.
}
\concept{GoogleCloudIdentitytoolkitAdminV2EmailTemplate functions}
