% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ForwardInboundCredentials}
\alias{GoogleCloudIdentitytoolkitAdminV2ForwardInboundCredentials}
\title{GoogleCloudIdentitytoolkitAdminV2ForwardInboundCredentials Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ForwardInboundCredentials(
  idToken = NULL,
  accessToken = NULL,
  refreshToken = NULL
)
}
\arguments{
\item{idToken}{Whether to pass the user's OIDC identity provider's ID token}

\item{accessToken}{Whether to pass the user's OAuth identity provider's access token}

\item{refreshToken}{Whether to pass the user's OAuth identity provider's refresh token}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ForwardInboundCredentials object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ForwardInboundCredentials Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indicates which credentials to pass to the registered Blocking Functions.
}
\concept{GoogleCloudIdentitytoolkitAdminV2ForwardInboundCredentials functions}
