% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2HashConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2HashConfig}
\title{GoogleCloudIdentitytoolkitAdminV2HashConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2HashConfig()
}
\value{
GoogleCloudIdentitytoolkitAdminV2HashConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2HashConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
History information of the hash algorithm and key. Different accounts' passwords may be generated by different version.
}
\concept{GoogleCloudIdentitytoolkitAdminV2HashConfig functions}
