% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2IdpCertificate}
\alias{GoogleCloudIdentitytoolkitAdminV2IdpCertificate}
\title{GoogleCloudIdentitytoolkitAdminV2IdpCertificate Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2IdpCertificate(x509Certificate = NULL)
}
\arguments{
\item{x509Certificate}{The x509 certificate}
}
\value{
GoogleCloudIdentitytoolkitAdminV2IdpCertificate object
}
\description{
GoogleCloudIdentitytoolkitAdminV2IdpCertificate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The IDP's certificate data to verify the signature in the SAMLResponse issued by the IDP.
}
\concept{GoogleCloudIdentitytoolkitAdminV2IdpCertificate functions}
