% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2IdpConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2IdpConfig}
\title{GoogleCloudIdentitytoolkitAdminV2IdpConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2IdpConfig(
  signRequest = NULL,
  idpCertificates = NULL,
  idpEntityId = NULL,
  ssoUrl = NULL
)
}
\arguments{
\item{signRequest}{Indicates if outbounding SAMLRequest should be signed}

\item{idpCertificates}{IDP's public keys for verifying signature in the assertions}

\item{idpEntityId}{Unique identifier for all SAML entities}

\item{ssoUrl}{URL to send Authentication request to}
}
\value{
GoogleCloudIdentitytoolkitAdminV2IdpConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2IdpConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The SAML IdP (Identity Provider) configuration when the project acts as the relying party.
}
\concept{GoogleCloudIdentitytoolkitAdminV2IdpConfig functions}
