% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig}
\title{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig(
  idpConfig = NULL,
  displayName = NULL,
  spConfig = NULL,
  name = NULL,
  enabled = NULL
)
}
\arguments{
\item{idpConfig}{The SAML IdP (Identity Provider) configuration when the project acts as the relying party}

\item{displayName}{The config's display name set by developers}

\item{spConfig}{The SAML SP (Service Provider) configuration when the project acts as the relying party to receive and accept an authentication assertion issued by a SAML identity provider}

\item{name}{The name of the InboundSamlConfig resource, for example: 'projects/my-awesome-project/inboundSamlConfigs/my-config-id'}

\item{enabled}{True if allows the user to sign in with the provider}
}
\value{
GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A pair of SAML RP-IDP configurations when the project acts as the relying party.
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig functions: 
\code{\link{projects.inboundSamlConfigs.create}()},
\code{\link{projects.inboundSamlConfigs.patch}()},
\code{\link{projects.tenants.inboundSamlConfigs.create}()},
\code{\link{projects.tenants.inboundSamlConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig functions}
