% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Inheritance}
\alias{GoogleCloudIdentitytoolkitAdminV2Inheritance}
\title{GoogleCloudIdentitytoolkitAdminV2Inheritance Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Inheritance(emailSendingConfig = NULL)
}
\arguments{
\item{emailSendingConfig}{Whether to allow the tenant to inherit custom domains, email templates, and custom SMTP settings}
}
\value{
GoogleCloudIdentitytoolkitAdminV2Inheritance object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Inheritance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings that the tenants will inherit from project level.
}
\concept{GoogleCloudIdentitytoolkitAdminV2Inheritance functions}
