% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpConfigsResponse}
\alias{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpConfigsResponse}
\title{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpConfigsResponse Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpConfigsResponse(
  nextPageToken = NULL,
  defaultSupportedIdpConfigs = NULL
)
}
\arguments{
\item{nextPageToken}{Token to retrieve the next page of results, or empty if there are no more results in the list}

\item{defaultSupportedIdpConfigs}{The set of configs}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpConfigsResponse object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpConfigsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for DefaultSupportedIdpConfigs
}
\concept{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpConfigsResponse functions}
