% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpsResponse}
\alias{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpsResponse}
\title{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpsResponse Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpsResponse(
  defaultSupportedIdps = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{defaultSupportedIdps}{The set of configs}

\item{nextPageToken}{Token to retrieve the next page of results, or empty if there are no more results in the list}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpsResponse object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for ListDefaultSupportedIdps
}
\concept{GoogleCloudIdentitytoolkitAdminV2ListDefaultSupportedIdpsResponse functions}
