% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ListInboundSamlConfigsResponse}
\alias{GoogleCloudIdentitytoolkitAdminV2ListInboundSamlConfigsResponse}
\title{GoogleCloudIdentitytoolkitAdminV2ListInboundSamlConfigsResponse Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ListInboundSamlConfigsResponse(
  nextPageToken = NULL,
  inboundSamlConfigs = NULL
)
}
\arguments{
\item{nextPageToken}{Token to retrieve the next page of results, or empty if there are no more results in the list}

\item{inboundSamlConfigs}{The set of configs}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ListInboundSamlConfigsResponse object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ListInboundSamlConfigsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for ListInboundSamlConfigs
}
\concept{GoogleCloudIdentitytoolkitAdminV2ListInboundSamlConfigsResponse functions}
