% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ListOAuthIdpConfigsResponse}
\alias{GoogleCloudIdentitytoolkitAdminV2ListOAuthIdpConfigsResponse}
\title{GoogleCloudIdentitytoolkitAdminV2ListOAuthIdpConfigsResponse Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ListOAuthIdpConfigsResponse(
  oauthIdpConfigs = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{oauthIdpConfigs}{The set of configs}

\item{nextPageToken}{Token to retrieve the next page of results, or empty if there are no more results in the list}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ListOAuthIdpConfigsResponse object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ListOAuthIdpConfigsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for ListOAuthIdpConfigs
}
\concept{GoogleCloudIdentitytoolkitAdminV2ListOAuthIdpConfigsResponse functions}
