% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ListTenantsResponse}
\alias{GoogleCloudIdentitytoolkitAdminV2ListTenantsResponse}
\title{GoogleCloudIdentitytoolkitAdminV2ListTenantsResponse Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ListTenantsResponse(
  nextPageToken = NULL,
  tenants = NULL
)
}
\arguments{
\item{nextPageToken}{The token to get the next page of results}

\item{tenants}{A list of tenants under the given agent project}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ListTenantsResponse object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ListTenantsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ListTenants.
}
\concept{GoogleCloudIdentitytoolkitAdminV2ListTenantsResponse functions}
