% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2MonitoringConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2MonitoringConfig}
\title{GoogleCloudIdentitytoolkitAdminV2MonitoringConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2MonitoringConfig(requestLogging = NULL)
}
\arguments{
\item{requestLogging}{Configuration for logging requests made to this project to Stackdriver Logging}
}
\value{
GoogleCloudIdentitytoolkitAdminV2MonitoringConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2MonitoringConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration related to monitoring project activity.
}
\concept{GoogleCloudIdentitytoolkitAdminV2MonitoringConfig functions}
