% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2MultiFactorAuthConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2MultiFactorAuthConfig}
\title{GoogleCloudIdentitytoolkitAdminV2MultiFactorAuthConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2MultiFactorAuthConfig(
  state = NULL,
  providerConfigs = NULL,
  enabledProviders = NULL
)
}
\arguments{
\item{state}{Whether MultiFactor Authentication has been enabled for this project}

\item{providerConfigs}{A list of usable second factors for this project along with their configurations}

\item{enabledProviders}{A list of usable second factors for this project}
}
\value{
GoogleCloudIdentitytoolkitAdminV2MultiFactorAuthConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2MultiFactorAuthConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options related to MultiFactor Authentication for the project.
}
\concept{GoogleCloudIdentitytoolkitAdminV2MultiFactorAuthConfig functions}
