% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2MultiTenantConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2MultiTenantConfig}
\title{GoogleCloudIdentitytoolkitAdminV2MultiTenantConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2MultiTenantConfig(
  defaultTenantLocation = NULL,
  allowTenants = NULL
)
}
\arguments{
\item{defaultTenantLocation}{The default cloud parent org or folder that the tenant project should be created under}

\item{allowTenants}{Whether this project can have tenants or not}
}
\value{
GoogleCloudIdentitytoolkitAdminV2MultiTenantConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2MultiTenantConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration related to multi-tenant functionality.
}
\concept{GoogleCloudIdentitytoolkitAdminV2MultiTenantConfig functions}
