% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2NotificationConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2NotificationConfig}
\title{GoogleCloudIdentitytoolkitAdminV2NotificationConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2NotificationConfig(
  sendEmail = NULL,
  sendSms = NULL,
  defaultLocale = NULL
)
}
\arguments{
\item{sendEmail}{Options for email sending}

\item{sendSms}{Options for SMS sending}

\item{defaultLocale}{Default locale used for email and SMS in IETF BCP 47 format}
}
\value{
GoogleCloudIdentitytoolkitAdminV2NotificationConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2NotificationConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration related to sending notifications to users.
}
\concept{GoogleCloudIdentitytoolkitAdminV2NotificationConfig functions}
