% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig}
\title{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig(
  name = NULL,
  clientSecret = NULL,
  clientId = NULL,
  issuer = NULL,
  enabled = NULL,
  responseType = NULL,
  displayName = NULL
)
}
\arguments{
\item{name}{The name of the OAuthIdpConfig resource, for example: 'projects/my-awesome-project/oauthIdpConfigs/oauth-config-id'}

\item{clientSecret}{The client secret of the OAuth client, to enable OIDC code flow}

\item{clientId}{The client id of an OAuth client}

\item{issuer}{For OIDC Idps, the issuer identifier}

\item{enabled}{True if allows the user to sign in with the provider}

\item{responseType}{The response type to request for in the OAuth authorization flow}

\item{displayName}{The config's display name set by developers}
}
\value{
GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration options for authenticating with an OAuth IDP.
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig functions: 
\code{\link{projects.oauthIdpConfigs.create}()},
\code{\link{projects.oauthIdpConfigs.patch}()},
\code{\link{projects.tenants.oauthIdpConfigs.create}()},
\code{\link{projects.tenants.oauthIdpConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig functions}
