% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2OAuthResponseType}
\alias{GoogleCloudIdentitytoolkitAdminV2OAuthResponseType}
\title{GoogleCloudIdentitytoolkitAdminV2OAuthResponseType Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2OAuthResponseType(
  idToken = NULL,
  code = NULL,
  token = NULL
)
}
\arguments{
\item{idToken}{If true, ID token is returned from IdP's authorization endpoint}

\item{code}{If true, authorization code is returned from IdP's authorization endpoint}

\item{token}{Do not use}
}
\value{
GoogleCloudIdentitytoolkitAdminV2OAuthResponseType object
}
\description{
GoogleCloudIdentitytoolkitAdminV2OAuthResponseType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response type to request for in the OAuth authorization flow. You can set either \code{id_token} or \code{code} to true, but not both. Setting both types to be simultaneously true (\verb{\{code: true, id_token: true\}}) is not yet supported. See https://openid.net/specs/openid-connect-core-1_0.html#Authentication for a mapping of response type to OAuth 2.0 flow.
}
\concept{GoogleCloudIdentitytoolkitAdminV2OAuthResponseType functions}
