% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2PasswordPolicyConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2PasswordPolicyConfig}
\title{Identity Toolkit API Objects
The Google Identity Toolkit API lets you use open standards to verify a user's identity.}
\usage{
GoogleCloudIdentitytoolkitAdminV2PasswordPolicyConfig(
  passwordPolicyVersions = NULL,
  passwordPolicyEnforcementState = NULL,
  forceUpgradeOnSignin = NULL
)
}
\arguments{
\item{passwordPolicyVersions}{Must be of length 1}

\item{passwordPolicyEnforcementState}{Which enforcement mode to use for the password policy}

\item{forceUpgradeOnSignin}{Users must have a password compliant with the password policy to sign-in}
}
\value{
GoogleCloudIdentitytoolkitAdminV2PasswordPolicyConfig object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2023-06-21 08:48:24.872748
filename: /Users/justinjm/projects/github.com/justinjm/autoGoogleAPI/googleidentitytoolkitv2.auto/R/identitytoolkit_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GoogleCloudIdentitytoolkitAdminV2PasswordPolicyConfig Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The configuration for the password policy on the project.
}
\concept{GoogleCloudIdentitytoolkitAdminV2PasswordPolicyConfig functions}
