% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2PasswordPolicyVersion}
\alias{GoogleCloudIdentitytoolkitAdminV2PasswordPolicyVersion}
\title{GoogleCloudIdentitytoolkitAdminV2PasswordPolicyVersion Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2PasswordPolicyVersion(
  customStrengthOptions = NULL
)
}
\arguments{
\item{customStrengthOptions}{The custom strength options enforced by the password policy}
}
\value{
GoogleCloudIdentitytoolkitAdminV2PasswordPolicyVersion object
}
\description{
GoogleCloudIdentitytoolkitAdminV2PasswordPolicyVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The strength attributes for the password policy on the project.
}
\concept{GoogleCloudIdentitytoolkitAdminV2PasswordPolicyVersion functions}
