% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2PhoneNumber}
\alias{GoogleCloudIdentitytoolkitAdminV2PhoneNumber}
\title{GoogleCloudIdentitytoolkitAdminV2PhoneNumber Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2PhoneNumber(
  GoogleCloudIdentitytoolkitAdminV2PhoneNumber.testPhoneNumbers = NULL,
  enabled = NULL,
  testPhoneNumbers = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2PhoneNumber.testPhoneNumbers}{The \link{GoogleCloudIdentitytoolkitAdminV2PhoneNumber.testPhoneNumbers} object or list of objects}

\item{enabled}{Whether phone number auth is enabled for the project or not}

\item{testPhoneNumbers}{A map of that can be used for phone auth testing}
}
\value{
GoogleCloudIdentitytoolkitAdminV2PhoneNumber object
}
\description{
GoogleCloudIdentitytoolkitAdminV2PhoneNumber Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration options related to authenticated a user by their phone number.
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2PhoneNumber functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2PhoneNumber.testPhoneNumbers}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2PhoneNumber functions}
