% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2ProviderConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2ProviderConfig}
\title{GoogleCloudIdentitytoolkitAdminV2ProviderConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2ProviderConfig(
  totpProviderConfig = NULL,
  state = NULL
)
}
\arguments{
\item{totpProviderConfig}{TOTP MFA provider config for this project}

\item{state}{Describes the state of the MultiFactor Authentication type}
}
\value{
GoogleCloudIdentitytoolkitAdminV2ProviderConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2ProviderConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ProviderConfig describes the supported MFA providers along with their configurations.
}
\concept{GoogleCloudIdentitytoolkitAdminV2ProviderConfig functions}
