% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2QuotaConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2QuotaConfig}
\title{GoogleCloudIdentitytoolkitAdminV2QuotaConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2QuotaConfig(signUpQuotaConfig = NULL)
}
\arguments{
\item{signUpQuotaConfig}{Quota for the Signup endpoint, if overwritten}
}
\value{
GoogleCloudIdentitytoolkitAdminV2QuotaConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2QuotaConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration related to quotas.
}
\concept{GoogleCloudIdentitytoolkitAdminV2QuotaConfig functions}
