% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2RecaptchaConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2RecaptchaConfig}
\title{GoogleCloudIdentitytoolkitAdminV2RecaptchaConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2RecaptchaConfig(
  useAccountDefender = NULL,
  managedRules = NULL,
  emailPasswordEnforcementState = NULL
)
}
\arguments{
\item{useAccountDefender}{Whether to use the account defender for reCAPTCHA assessment}

\item{managedRules}{The managed rules for authentication action based on reCAPTCHA scores}

\item{emailPasswordEnforcementState}{The reCAPTCHA config for email/password provider, containing the enforcement status}
}
\value{
GoogleCloudIdentitytoolkitAdminV2RecaptchaConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2RecaptchaConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The reCAPTCHA Enterprise integration config.
}
\concept{GoogleCloudIdentitytoolkitAdminV2RecaptchaConfig functions}
