% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2RecaptchaKey}
\alias{GoogleCloudIdentitytoolkitAdminV2RecaptchaKey}
\title{GoogleCloudIdentitytoolkitAdminV2RecaptchaKey Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2RecaptchaKey(type = NULL, key = NULL)
}
\arguments{
\item{type}{The client's platform type}

\item{key}{The reCAPTCHA Enterprise key resource name, e}
}
\value{
GoogleCloudIdentitytoolkitAdminV2RecaptchaKey object
}
\description{
GoogleCloudIdentitytoolkitAdminV2RecaptchaKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The reCAPTCHA key config. reCAPTCHA Enterprise offers different keys for different client platforms.
}
\concept{GoogleCloudIdentitytoolkitAdminV2RecaptchaKey functions}
