% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2RecaptchaManagedRule}
\alias{GoogleCloudIdentitytoolkitAdminV2RecaptchaManagedRule}
\title{GoogleCloudIdentitytoolkitAdminV2RecaptchaManagedRule Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2RecaptchaManagedRule(
  action = NULL,
  endScore = NULL
)
}
\arguments{
\item{action}{The action taken if the reCAPTCHA score of a request is within the interval \link{start_score, end_score}}

\item{endScore}{The end score (inclusive) of the score range for an action}
}
\value{
GoogleCloudIdentitytoolkitAdminV2RecaptchaManagedRule object
}
\description{
GoogleCloudIdentitytoolkitAdminV2RecaptchaManagedRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config for a reCAPTCHA managed rule. Models a single interval \link{start_score, end_score}. The start_score is implicit. It is either the closest smaller end_score (if one is available) or 0. Intervals in aggregate span \link{0, 1} without overlapping.
}
\concept{GoogleCloudIdentitytoolkitAdminV2RecaptchaManagedRule functions}
