% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2RequestLogging}
\alias{GoogleCloudIdentitytoolkitAdminV2RequestLogging}
\title{GoogleCloudIdentitytoolkitAdminV2RequestLogging Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2RequestLogging(enabled = NULL)
}
\arguments{
\item{enabled}{Whether logging is enabled for this project or not}
}
\value{
GoogleCloudIdentitytoolkitAdminV2RequestLogging object
}
\description{
GoogleCloudIdentitytoolkitAdminV2RequestLogging Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for logging requests made to this project to Stackdriver Logging
}
\concept{GoogleCloudIdentitytoolkitAdminV2RequestLogging functions}
