% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2SendEmail}
\alias{GoogleCloudIdentitytoolkitAdminV2SendEmail}
\title{GoogleCloudIdentitytoolkitAdminV2SendEmail Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2SendEmail(
  smtp = NULL,
  changeEmailTemplate = NULL,
  method = NULL,
  dnsInfo = NULL,
  legacyResetPasswordTemplate = NULL,
  callbackUri = NULL,
  revertSecondFactorAdditionTemplate = NULL,
  resetPasswordTemplate = NULL,
  verifyEmailTemplate = NULL
)
}
\arguments{
\item{smtp}{Use a custom SMTP relay}

\item{changeEmailTemplate}{Email template for change email}

\item{method}{The method used for sending an email}

\item{dnsInfo}{Information of custom domain DNS verification}

\item{legacyResetPasswordTemplate}{Reset password email template for legacy Firebase V1 app}

\item{callbackUri}{action url in email template}

\item{revertSecondFactorAdditionTemplate}{Email template for reverting second factor addition emails}

\item{resetPasswordTemplate}{Email template for reset password}

\item{verifyEmailTemplate}{Email template for verify email}
}
\value{
GoogleCloudIdentitytoolkitAdminV2SendEmail object
}
\description{
GoogleCloudIdentitytoolkitAdminV2SendEmail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for email sending.
}
\concept{GoogleCloudIdentitytoolkitAdminV2SendEmail functions}
