% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2SendSms}
\alias{GoogleCloudIdentitytoolkitAdminV2SendSms}
\title{GoogleCloudIdentitytoolkitAdminV2SendSms Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2SendSms(useDeviceLocale = NULL)
}
\arguments{
\item{useDeviceLocale}{Whether to use the accept_language header for SMS}
}
\value{
GoogleCloudIdentitytoolkitAdminV2SendSms object
}
\description{
GoogleCloudIdentitytoolkitAdminV2SendSms Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for SMS sending.
}
\concept{GoogleCloudIdentitytoolkitAdminV2SendSms functions}
