% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2SignInConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2SignInConfig}
\title{GoogleCloudIdentitytoolkitAdminV2SignInConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2SignInConfig(
  email = NULL,
  allowDuplicateEmails = NULL,
  phoneNumber = NULL,
  anonymous = NULL
)
}
\arguments{
\item{email}{Configuration options related to authenticating a user by their email address}

\item{allowDuplicateEmails}{Whether to allow more than one account to have the same email}

\item{phoneNumber}{Configuration options related to authenticated a user by their phone number}

\item{anonymous}{Configuration options related to authenticating an anonymous user}
}
\value{
GoogleCloudIdentitytoolkitAdminV2SignInConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2SignInConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration related to local sign in methods.
}
\concept{GoogleCloudIdentitytoolkitAdminV2SignInConfig functions}
