% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2SmsRegionConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2SmsRegionConfig}
\title{GoogleCloudIdentitytoolkitAdminV2SmsRegionConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2SmsRegionConfig(
  allowByDefault = NULL,
  allowlistOnly = NULL
)
}
\arguments{
\item{allowByDefault}{A policy of allowing SMS to every region by default and adding disallowed regions to a disallow list}

\item{allowlistOnly}{A policy of only allowing regions by explicitly adding them to an allowlist}
}
\value{
GoogleCloudIdentitytoolkitAdminV2SmsRegionConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2SmsRegionConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configures the regions where users are allowed to send verification SMS for the project or tenant. This is based on the calling code of the destination phone number.
}
\concept{GoogleCloudIdentitytoolkitAdminV2SmsRegionConfig functions}
