% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2SmsTemplate}
\alias{GoogleCloudIdentitytoolkitAdminV2SmsTemplate}
\title{GoogleCloudIdentitytoolkitAdminV2SmsTemplate Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2SmsTemplate()
}
\value{
GoogleCloudIdentitytoolkitAdminV2SmsTemplate object
}
\description{
GoogleCloudIdentitytoolkitAdminV2SmsTemplate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The template to use when sending an SMS.
}
\concept{GoogleCloudIdentitytoolkitAdminV2SmsTemplate functions}
