% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Smtp}
\alias{GoogleCloudIdentitytoolkitAdminV2Smtp}
\title{GoogleCloudIdentitytoolkitAdminV2Smtp Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Smtp(
  password = NULL,
  username = NULL,
  port = NULL,
  securityMode = NULL,
  host = NULL,
  senderEmail = NULL
)
}
\arguments{
\item{password}{SMTP relay password}

\item{username}{SMTP relay username}

\item{port}{SMTP relay port}

\item{securityMode}{SMTP security mode}

\item{host}{SMTP relay host}

\item{senderEmail}{Sender email for the SMTP relay}
}
\value{
GoogleCloudIdentitytoolkitAdminV2Smtp object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Smtp Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for SMTP relay
}
\concept{GoogleCloudIdentitytoolkitAdminV2Smtp functions}
