% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2SpCertificate}
\alias{GoogleCloudIdentitytoolkitAdminV2SpCertificate}
\title{GoogleCloudIdentitytoolkitAdminV2SpCertificate Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2SpCertificate(
  x509Certificate = NULL,
  expiresAt = NULL
)
}
\arguments{
\item{x509Certificate}{Self-signed public certificate}

\item{expiresAt}{Timestamp of the cert expiration instance}
}
\value{
GoogleCloudIdentitytoolkitAdminV2SpCertificate object
}
\description{
GoogleCloudIdentitytoolkitAdminV2SpCertificate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The SP's certificate data for IDP to verify the SAMLRequest generated by the SP.
}
\concept{GoogleCloudIdentitytoolkitAdminV2SpCertificate functions}
