% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2SpConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2SpConfig}
\title{GoogleCloudIdentitytoolkitAdminV2SpConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2SpConfig(
  spEntityId = NULL,
  callbackUri = NULL
)
}
\arguments{
\item{spEntityId}{Unique identifier for all SAML entities}

\item{callbackUri}{Callback URI where responses from IDP are handled}
}
\value{
GoogleCloudIdentitytoolkitAdminV2SpConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2SpConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The SAML SP (Service Provider) configuration when the project acts as the relying party to receive and accept an authentication assertion issued by a SAML identity provider.
}
\concept{GoogleCloudIdentitytoolkitAdminV2SpConfig functions}
