% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2TemporaryQuota}
\alias{GoogleCloudIdentitytoolkitAdminV2TemporaryQuota}
\title{GoogleCloudIdentitytoolkitAdminV2TemporaryQuota Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2TemporaryQuota(
  quota = NULL,
  quotaDuration = NULL,
  startTime = NULL
)
}
\arguments{
\item{quota}{Corresponds to the 'refill_token_count' field in QuotaServer config}

\item{quotaDuration}{How long this quota will be active for}

\item{startTime}{When this quota will take affect}
}
\value{
GoogleCloudIdentitytoolkitAdminV2TemporaryQuota object
}
\description{
GoogleCloudIdentitytoolkitAdminV2TemporaryQuota Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Temporary quota increase / decrease
}
\concept{GoogleCloudIdentitytoolkitAdminV2TemporaryQuota functions}
