% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Tenant}
\alias{GoogleCloudIdentitytoolkitAdminV2Tenant}
\title{GoogleCloudIdentitytoolkitAdminV2Tenant Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Tenant(
  GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers = NULL,
  monitoring = NULL,
  allowPasswordSignup = NULL,
  mfaConfig = NULL,
  passwordPolicyConfig = NULL,
  disableAuth = NULL,
  autodeleteAnonymousUsers = NULL,
  smsRegionConfig = NULL,
  displayName = NULL,
  client = NULL,
  testPhoneNumbers = NULL,
  inheritance = NULL,
  enableEmailLinkSignin = NULL,
  emailPrivacyConfig = NULL,
  enableAnonymousUser = NULL,
  recaptchaConfig = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers}{The \link{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers} object or list of objects}

\item{monitoring}{Configuration related to monitoring project activity}

\item{allowPasswordSignup}{Whether to allow email/password user authentication}

\item{mfaConfig}{The tenant-level configuration of MFA options}

\item{passwordPolicyConfig}{The tenant-level password policy config}

\item{disableAuth}{Whether authentication is disabled for the tenant}

\item{autodeleteAnonymousUsers}{Whether anonymous users will be auto-deleted after a period of 30 days}

\item{smsRegionConfig}{Configures which regions are enabled for SMS verification code sending}

\item{displayName}{Display name of the tenant}

\item{client}{Options related to how clients making requests on behalf of a project should be configured}

\item{testPhoneNumbers}{A map of pairs that can be used for MFA}

\item{inheritance}{Specify the settings that the tenant could inherit}

\item{enableEmailLinkSignin}{Whether to enable email link user authentication}

\item{emailPrivacyConfig}{Configuration for settings related to email privacy and public visibility}

\item{enableAnonymousUser}{Whether to enable anonymous user authentication}

\item{recaptchaConfig}{The tenant-level reCAPTCHA config}
}
\value{
GoogleCloudIdentitytoolkitAdminV2Tenant object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Tenant Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Tenant contains configuration for the tenant in a multi-tenant project.
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2Tenant functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers}()},
\code{\link{projects.tenants.create}()},
\code{\link{projects.tenants.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2Tenant functions}
