% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers}
\alias{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers}
\title{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers()
}
\value{
GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of pairs that can be used for MFA. The phone number should be in E.164 format (https://www.itu.int/rec/T-REC-E.164/) and a maximum of 10 pairs can be added (error will be thrown once exceeded).
}
\seealso{
Other GoogleCloudIdentitytoolkitAdminV2Tenant functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2Tenant}()},
\code{\link{projects.tenants.create}()},
\code{\link{projects.tenants.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2Tenant functions}
