% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2TotpMfaProviderConfig}
\alias{GoogleCloudIdentitytoolkitAdminV2TotpMfaProviderConfig}
\title{GoogleCloudIdentitytoolkitAdminV2TotpMfaProviderConfig Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2TotpMfaProviderConfig(
  adjacentIntervals = NULL
)
}
\arguments{
\item{adjacentIntervals}{The allowed number of adjacent intervals that will be used for verification to avoid clock skew}
}
\value{
GoogleCloudIdentitytoolkitAdminV2TotpMfaProviderConfig object
}
\description{
GoogleCloudIdentitytoolkitAdminV2TotpMfaProviderConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TotpMFAProviderConfig represents the TOTP based MFA provider.
}
\concept{GoogleCloudIdentitytoolkitAdminV2TotpMfaProviderConfig functions}
