% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitAdminV2Trigger}
\alias{GoogleCloudIdentitytoolkitAdminV2Trigger}
\title{GoogleCloudIdentitytoolkitAdminV2Trigger Object}
\usage{
GoogleCloudIdentitytoolkitAdminV2Trigger(updateTime = NULL, functionUri = NULL)
}
\arguments{
\item{updateTime}{When the trigger was changed}

\item{functionUri}{HTTP URI trigger for the Cloud Function}
}
\value{
GoogleCloudIdentitytoolkitAdminV2Trigger object
}
\description{
GoogleCloudIdentitytoolkitAdminV2Trigger Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Synchronous Cloud Function with HTTP Trigger
}
\concept{GoogleCloudIdentitytoolkitAdminV2Trigger functions}
