% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2AutoRetrievalInfo}
\alias{GoogleCloudIdentitytoolkitV2AutoRetrievalInfo}
\title{GoogleCloudIdentitytoolkitV2AutoRetrievalInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2AutoRetrievalInfo(appSignatureHash = NULL)
}
\arguments{
\item{appSignatureHash}{The Android app's signature hash for Google Play Service's SMS Retriever API}
}
\value{
GoogleCloudIdentitytoolkitV2AutoRetrievalInfo object
}
\description{
GoogleCloudIdentitytoolkitV2AutoRetrievalInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The information required to auto-retrieve an SMS.
}
\concept{GoogleCloudIdentitytoolkitV2AutoRetrievalInfo functions}
