% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2CustomStrengthOptions}
\alias{GoogleCloudIdentitytoolkitV2CustomStrengthOptions}
\title{GoogleCloudIdentitytoolkitV2CustomStrengthOptions Object}
\usage{
GoogleCloudIdentitytoolkitV2CustomStrengthOptions(
  containsNonAlphanumericCharacter = NULL,
  maxPasswordLength = NULL,
  containsLowercaseCharacter = NULL,
  containsUppercaseCharacter = NULL,
  containsNumericCharacter = NULL,
  minPasswordLength = NULL
)
}
\arguments{
\item{containsNonAlphanumericCharacter}{The password must contain a non alpha numeric character}

\item{maxPasswordLength}{Maximum password length}

\item{containsLowercaseCharacter}{The password must contain a lower case character}

\item{containsUppercaseCharacter}{The password must contain an upper case character}

\item{containsNumericCharacter}{The password must contain a number}

\item{minPasswordLength}{Minimum password length}
}
\value{
GoogleCloudIdentitytoolkitV2CustomStrengthOptions object
}
\description{
GoogleCloudIdentitytoolkitV2CustomStrengthOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Custom strength options to enforce on user passwords.
}
\concept{GoogleCloudIdentitytoolkitV2CustomStrengthOptions functions}
