% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest}
\alias{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest}
\title{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest Object}
\usage{
GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest(
  phoneVerificationInfo = NULL,
  totpVerificationInfo = NULL,
  displayName = NULL,
  tenantId = NULL,
  idToken = NULL
)
}
\arguments{
\item{phoneVerificationInfo}{Verification info to authorize sending an SMS for phone verification}

\item{totpVerificationInfo}{Verification information for TOTP}

\item{displayName}{Display name which is entered by users to distinguish between different second factors with same type or different type}

\item{tenantId}{The ID of the Identity Platform tenant that the user enrolling MFA belongs to}

\item{idToken}{Required}
}
\value{
GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest object
}
\description{
GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Finishes enrolling a second factor for the user.
}
\seealso{
Other GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest functions: 
\code{\link{accounts.mfaEnrollment.finalize}()}
}
\concept{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentRequest functions}
