% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentResponse}
\alias{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentResponse}
\title{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentResponse Object}
\usage{
GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentResponse(
  phoneAuthInfo = NULL,
  refreshToken = NULL,
  idToken = NULL,
  totpAuthInfo = NULL
)
}
\arguments{
\item{phoneAuthInfo}{Auxiliary auth info specific to phone auth}

\item{refreshToken}{Refresh token updated to reflect MFA enrollment}

\item{idToken}{ID token updated to reflect MFA enrollment}

\item{totpAuthInfo}{Auxiliary auth info specific to TOTP auth}
}
\value{
GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentResponse object
}
\description{
GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
FinalizeMfaEnrollment response.
}
\concept{GoogleCloudIdentitytoolkitV2FinalizeMfaEnrollmentResponse functions}
