% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneRequestInfo}
\alias{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneRequestInfo}
\title{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneRequestInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneRequestInfo(
  androidVerificationProof = NULL,
  code = NULL,
  phoneNumber = NULL,
  sessionInfo = NULL
)
}
\arguments{
\item{androidVerificationProof}{Android only}

\item{code}{User-entered verification code}

\item{phoneNumber}{Required if Android verification proof is presented}

\item{sessionInfo}{An opaque string that represents the enrollment session}
}
\value{
GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneRequestInfo object
}
\description{
GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneRequestInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Phone Verification info for a FinalizeMfa request.
}
\concept{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneRequestInfo functions}
