% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneResponseInfo}
\alias{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneResponseInfo}
\title{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneResponseInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneResponseInfo(
  androidVerificationProof = NULL,
  phoneNumber = NULL,
  androidVerificationProofExpireTime = NULL
)
}
\arguments{
\item{androidVerificationProof}{Android only}

\item{phoneNumber}{For Android verification proof}

\item{androidVerificationProofExpireTime}{Android only}
}
\value{
GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneResponseInfo object
}
\description{
GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneResponseInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Phone Verification info for a FinalizeMfa response.
}
\concept{GoogleCloudIdentitytoolkitV2FinalizeMfaPhoneResponseInfo functions}
