% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest}
\alias{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest}
\title{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest Object}
\usage{
GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest(
  totpVerificationInfo = NULL,
  tenantId = NULL,
  mfaEnrollmentId = NULL,
  mfaPendingCredential = NULL,
  phoneVerificationInfo = NULL
)
}
\arguments{
\item{totpVerificationInfo}{Proof of completion of the TOTP based MFA challenge}

\item{tenantId}{The ID of the Identity Platform tenant the user is signing in to}

\item{mfaEnrollmentId}{The MFA enrollment ID from the user's list of current MFA enrollments}

\item{mfaPendingCredential}{Required}

\item{phoneVerificationInfo}{Proof of completion of the SMS based MFA challenge}
}
\value{
GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest object
}
\description{
GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Finalizes sign-in by verifying MFA challenge.
}
\seealso{
Other GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest functions: 
\code{\link{accounts.mfaSignIn.finalize}()}
}
\concept{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInRequest functions}
