% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInResponse}
\alias{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInResponse}
\title{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInResponse Object}
\usage{
GoogleCloudIdentitytoolkitV2FinalizeMfaSignInResponse(
  refreshToken = NULL,
  phoneAuthInfo = NULL,
  idToken = NULL
)
}
\arguments{
\item{refreshToken}{Refresh token for the authenticated user}

\item{phoneAuthInfo}{Extra phone auth info, including android verification proof}

\item{idToken}{ID token for the authenticated user}
}
\value{
GoogleCloudIdentitytoolkitV2FinalizeMfaSignInResponse object
}
\description{
GoogleCloudIdentitytoolkitV2FinalizeMfaSignInResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
FinalizeMfaSignIn response.
}
\concept{GoogleCloudIdentitytoolkitV2FinalizeMfaSignInResponse functions}
