% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2FinalizeMfaTotpEnrollmentRequestInfo}
\alias{GoogleCloudIdentitytoolkitV2FinalizeMfaTotpEnrollmentRequestInfo}
\title{GoogleCloudIdentitytoolkitV2FinalizeMfaTotpEnrollmentRequestInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2FinalizeMfaTotpEnrollmentRequestInfo(
  verificationCode = NULL,
  sessionInfo = NULL
)
}
\arguments{
\item{verificationCode}{User-entered verification code}

\item{sessionInfo}{An opaque string that represents the enrollment session}
}
\value{
GoogleCloudIdentitytoolkitV2FinalizeMfaTotpEnrollmentRequestInfo object
}
\description{
GoogleCloudIdentitytoolkitV2FinalizeMfaTotpEnrollmentRequestInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Mfa request info specific to TOTP auth for FinalizeMfa.
}
\concept{GoogleCloudIdentitytoolkitV2FinalizeMfaTotpEnrollmentRequestInfo functions}
