% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2MfaTotpSignInRequestInfo}
\alias{GoogleCloudIdentitytoolkitV2MfaTotpSignInRequestInfo}
\title{GoogleCloudIdentitytoolkitV2MfaTotpSignInRequestInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2MfaTotpSignInRequestInfo(verificationCode = NULL)
}
\arguments{
\item{verificationCode}{User-entered verification code}
}
\value{
GoogleCloudIdentitytoolkitV2MfaTotpSignInRequestInfo object
}
\description{
GoogleCloudIdentitytoolkitV2MfaTotpSignInRequestInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TOTP verification info for FinalizeMfaSignInRequest.
}
\concept{GoogleCloudIdentitytoolkitV2MfaTotpSignInRequestInfo functions}
