% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2PasswordPolicy}
\alias{GoogleCloudIdentitytoolkitV2PasswordPolicy}
\title{GoogleCloudIdentitytoolkitV2PasswordPolicy Object}
\usage{
GoogleCloudIdentitytoolkitV2PasswordPolicy(customStrengthOptions = NULL)
}
\arguments{
\item{customStrengthOptions}{The custom strength options enforced by the password policy}
}
\value{
GoogleCloudIdentitytoolkitV2PasswordPolicy object
}
\description{
GoogleCloudIdentitytoolkitV2PasswordPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for password policy.
}
\concept{GoogleCloudIdentitytoolkitV2PasswordPolicy functions}
