% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2RecaptchaConfig}
\alias{GoogleCloudIdentitytoolkitV2RecaptchaConfig}
\title{GoogleCloudIdentitytoolkitV2RecaptchaConfig Object}
\usage{
GoogleCloudIdentitytoolkitV2RecaptchaConfig(
  recaptchaKey = NULL,
  recaptchaEnforcementState = NULL
)
}
\arguments{
\item{recaptchaKey}{The reCAPTCHA Enterprise key resource name, e}

\item{recaptchaEnforcementState}{The reCAPTCHA enforcement state for the providers that GCIP supports reCAPTCHA protection}
}
\value{
GoogleCloudIdentitytoolkitV2RecaptchaConfig object
}
\description{
GoogleCloudIdentitytoolkitV2RecaptchaConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for reCAPTCHA
}
\concept{GoogleCloudIdentitytoolkitV2RecaptchaConfig functions}
