% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2RecaptchaEnforcementState}
\alias{GoogleCloudIdentitytoolkitV2RecaptchaEnforcementState}
\title{GoogleCloudIdentitytoolkitV2RecaptchaEnforcementState Object}
\usage{
GoogleCloudIdentitytoolkitV2RecaptchaEnforcementState(
  enforcementState = NULL,
  provider = NULL
)
}
\arguments{
\item{enforcementState}{The reCAPTCHA enforcement state for the provider}

\item{provider}{The provider that has reCAPTCHA protection}
}
\value{
GoogleCloudIdentitytoolkitV2RecaptchaEnforcementState object
}
\description{
GoogleCloudIdentitytoolkitV2RecaptchaEnforcementState Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Enforcement states for reCAPTCHA protection.
}
\concept{GoogleCloudIdentitytoolkitV2RecaptchaEnforcementState functions}
