% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2RevokeTokenRequest}
\alias{GoogleCloudIdentitytoolkitV2RevokeTokenRequest}
\title{GoogleCloudIdentitytoolkitV2RevokeTokenRequest Object}
\usage{
GoogleCloudIdentitytoolkitV2RevokeTokenRequest(
  providerId = NULL,
  idToken = NULL,
  tokenType = NULL,
  redirectUri = NULL,
  tenantId = NULL,
  token = NULL
)
}
\arguments{
\item{providerId}{Required}

\item{idToken}{Required}

\item{tokenType}{Required}

\item{redirectUri}{The redirect URI provided in the initial authorization request made by the client to the IDP}

\item{tenantId}{The ID of the Identity Platform tenant the user is signing in to}

\item{token}{Required}
}
\value{
GoogleCloudIdentitytoolkitV2RevokeTokenRequest object
}
\description{
GoogleCloudIdentitytoolkitV2RevokeTokenRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for RevokeToken.
}
\seealso{
Other GoogleCloudIdentitytoolkitV2RevokeTokenRequest functions: 
\code{\link{accounts.revokeToken}()}
}
\concept{GoogleCloudIdentitytoolkitV2RevokeTokenRequest functions}
