% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest}
\alias{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest}
\title{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest Object}
\usage{
GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest(
  totpEnrollmentInfo = NULL,
  phoneEnrollmentInfo = NULL,
  tenantId = NULL,
  idToken = NULL
)
}
\arguments{
\item{totpEnrollmentInfo}{Sign-in info specific to TOTP auth}

\item{phoneEnrollmentInfo}{Verification info to authorize sending an SMS for phone verification}

\item{tenantId}{The ID of the Identity Platform tenant that the user enrolling MFA belongs to}

\item{idToken}{Required}
}
\value{
GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest object
}
\description{
GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sends MFA enrollment verification SMS for a user.
}
\seealso{
Other GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest functions: 
\code{\link{accounts.mfaEnrollment.start}()}
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentRequest functions}
