% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentResponse}
\alias{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentResponse}
\title{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentResponse Object}
\usage{
GoogleCloudIdentitytoolkitV2StartMfaEnrollmentResponse(
  phoneSessionInfo = NULL,
  totpSessionInfo = NULL
)
}
\arguments{
\item{phoneSessionInfo}{Verification info to authorize sending an SMS for phone verification}

\item{totpSessionInfo}{Enrollment response info specific to TOTP auth}
}
\value{
GoogleCloudIdentitytoolkitV2StartMfaEnrollmentResponse object
}
\description{
GoogleCloudIdentitytoolkitV2StartMfaEnrollmentResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
StartMfaEnrollment response.
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaEnrollmentResponse functions}
