% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2StartMfaPhoneRequestInfo}
\alias{GoogleCloudIdentitytoolkitV2StartMfaPhoneRequestInfo}
\title{GoogleCloudIdentitytoolkitV2StartMfaPhoneRequestInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2StartMfaPhoneRequestInfo(
  phoneNumber = NULL,
  iosReceipt = NULL,
  recaptchaToken = NULL,
  playIntegrityToken = NULL,
  iosSecret = NULL,
  safetyNetToken = NULL,
  autoRetrievalInfo = NULL
)
}
\arguments{
\item{phoneNumber}{Required for enrollment}

\item{iosReceipt}{iOS only}

\item{recaptchaToken}{Web only}

\item{playIntegrityToken}{Android only}

\item{iosSecret}{iOS only}

\item{safetyNetToken}{Android only}

\item{autoRetrievalInfo}{Android only}
}
\value{
GoogleCloudIdentitytoolkitV2StartMfaPhoneRequestInfo object
}
\description{
GoogleCloudIdentitytoolkitV2StartMfaPhoneRequestInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
App Verification info for a StartMfa request.
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaPhoneRequestInfo functions}
