% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2StartMfaPhoneResponseInfo}
\alias{GoogleCloudIdentitytoolkitV2StartMfaPhoneResponseInfo}
\title{GoogleCloudIdentitytoolkitV2StartMfaPhoneResponseInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2StartMfaPhoneResponseInfo(sessionInfo = NULL)
}
\arguments{
\item{sessionInfo}{An opaque string that represents the enrollment session}
}
\value{
GoogleCloudIdentitytoolkitV2StartMfaPhoneResponseInfo object
}
\description{
GoogleCloudIdentitytoolkitV2StartMfaPhoneResponseInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Phone Verification info for a StartMfa response.
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaPhoneResponseInfo functions}
