% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2StartMfaSignInRequest}
\alias{GoogleCloudIdentitytoolkitV2StartMfaSignInRequest}
\title{GoogleCloudIdentitytoolkitV2StartMfaSignInRequest Object}
\usage{
GoogleCloudIdentitytoolkitV2StartMfaSignInRequest(
  mfaEnrollmentId = NULL,
  tenantId = NULL,
  phoneSignInInfo = NULL,
  mfaPendingCredential = NULL
)
}
\arguments{
\item{mfaEnrollmentId}{Required}

\item{tenantId}{The ID of the Identity Platform tenant the user is signing in to}

\item{phoneSignInInfo}{Verification info to authorize sending an SMS for phone verification}

\item{mfaPendingCredential}{Required}
}
\value{
GoogleCloudIdentitytoolkitV2StartMfaSignInRequest object
}
\description{
GoogleCloudIdentitytoolkitV2StartMfaSignInRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Starts multi-factor sign-in by sending the multi-factor auth challenge.
}
\seealso{
Other GoogleCloudIdentitytoolkitV2StartMfaSignInRequest functions: 
\code{\link{accounts.mfaSignIn.start}()}
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaSignInRequest functions}
