% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2StartMfaSignInResponse}
\alias{GoogleCloudIdentitytoolkitV2StartMfaSignInResponse}
\title{GoogleCloudIdentitytoolkitV2StartMfaSignInResponse Object}
\usage{
GoogleCloudIdentitytoolkitV2StartMfaSignInResponse(phoneResponseInfo = NULL)
}
\arguments{
\item{phoneResponseInfo}{MultiFactor sign-in session information specific to SMS-type second factors}
}
\value{
GoogleCloudIdentitytoolkitV2StartMfaSignInResponse object
}
\description{
GoogleCloudIdentitytoolkitV2StartMfaSignInResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
StartMfaSignIn response.
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaSignInResponse functions}
