% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2StartMfaTotpEnrollmentResponseInfo}
\alias{GoogleCloudIdentitytoolkitV2StartMfaTotpEnrollmentResponseInfo}
\title{GoogleCloudIdentitytoolkitV2StartMfaTotpEnrollmentResponseInfo Object}
\usage{
GoogleCloudIdentitytoolkitV2StartMfaTotpEnrollmentResponseInfo(
  finalizeEnrollmentTime = NULL,
  periodSec = NULL,
  verificationCodeLength = NULL,
  sharedSecretKey = NULL,
  sessionInfo = NULL,
  hashingAlgorithm = NULL
)
}
\arguments{
\item{finalizeEnrollmentTime}{The time by which the enrollment must finish}

\item{periodSec}{Duration in seconds at which the verification code will change}

\item{verificationCodeLength}{The length of the verification code that needs to be generated}

\item{sharedSecretKey}{A base 32 encoded string that represents the shared TOTP secret}

\item{sessionInfo}{An encoded string that represents the enrollment session}

\item{hashingAlgorithm}{The hashing algorithm used to generate the verification code}
}
\value{
GoogleCloudIdentitytoolkitV2StartMfaTotpEnrollmentResponseInfo object
}
\description{
GoogleCloudIdentitytoolkitV2StartMfaTotpEnrollmentResponseInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Mfa response info specific to TOTP auth for StartMfa.
}
\concept{GoogleCloudIdentitytoolkitV2StartMfaTotpEnrollmentResponseInfo functions}
