% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleCloudIdentitytoolkitV2WithdrawMfaResponse}
\alias{GoogleCloudIdentitytoolkitV2WithdrawMfaResponse}
\title{GoogleCloudIdentitytoolkitV2WithdrawMfaResponse Object}
\usage{
GoogleCloudIdentitytoolkitV2WithdrawMfaResponse(
  refreshToken = NULL,
  idToken = NULL
)
}
\arguments{
\item{refreshToken}{Refresh token updated to reflect removal of the second factor}

\item{idToken}{ID token updated to reflect removal of the second factor}
}
\value{
GoogleCloudIdentitytoolkitV2WithdrawMfaResponse object
}
\description{
GoogleCloudIdentitytoolkitV2WithdrawMfaResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Withdraws MultiFactorAuth response.
}
\concept{GoogleCloudIdentitytoolkitV2WithdrawMfaResponse functions}
