% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleIamV1AuditConfig}
\alias{GoogleIamV1AuditConfig}
\title{GoogleIamV1AuditConfig Object}
\usage{
GoogleIamV1AuditConfig(service = NULL, auditLogConfigs = NULL)
}
\arguments{
\item{service}{Specifies a service that will be enabled for audit logging}

\item{auditLogConfigs}{The configuration for logging of each type of permission}
}
\value{
GoogleIamV1AuditConfig object
}
\description{
GoogleIamV1AuditConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs. If there are AuditConfigs for both \code{allServices} and a specific service, the union of the two AuditConfigs is used for that service: the log_types specified in each AuditConfig are enabled, and the exempted_members in each AuditLogConfig are exempted. Example Policy with multiple AuditConfigs: { 'audit_configs': [ { 'service': 'allServices', 'audit_log_configs': [ { 'log_type': 'DATA_READ', 'exempted_members': \link{ 'user:jose@example.com' } }, { 'log_type': 'DATA_WRITE' }, { 'log_type': 'ADMIN_READ' } ] }, { 'service': 'sampleservice.googleapis.com', 'audit_log_configs': [ { 'log_type': 'DATA_READ' }, { 'log_type': 'DATA_WRITE', 'exempted_members': \link{ 'user:aliya@example.com' } } ] } ] } For sampleservice, this policy enables DATA_READ, DATA_WRITE and ADMIN_READ logging. It also exempts \code{jose@example.com} from DATA_READ logging, and \code{aliya@example.com} from DATA_WRITE logging.
}
\concept{GoogleIamV1AuditConfig functions}
