% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleIamV1GetIamPolicyRequest}
\alias{GoogleIamV1GetIamPolicyRequest}
\title{GoogleIamV1GetIamPolicyRequest Object}
\usage{
GoogleIamV1GetIamPolicyRequest(options = NULL)
}
\arguments{
\item{options}{OPTIONAL: A \code{GetPolicyOptions} object for specifying options to \code{GetIamPolicy}}
}
\value{
GoogleIamV1GetIamPolicyRequest object
}
\description{
GoogleIamV1GetIamPolicyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for \code{GetIamPolicy} method.
}
\seealso{
Other GoogleIamV1GetIamPolicyRequest functions: 
\code{\link{projects.tenants.getIamPolicy}()}
}
\concept{GoogleIamV1GetIamPolicyRequest functions}
