% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleIamV1GetPolicyOptions}
\alias{GoogleIamV1GetPolicyOptions}
\title{GoogleIamV1GetPolicyOptions Object}
\usage{
GoogleIamV1GetPolicyOptions(requestedPolicyVersion = NULL)
}
\arguments{
\item{requestedPolicyVersion}{Optional}
}
\value{
GoogleIamV1GetPolicyOptions object
}
\description{
GoogleIamV1GetPolicyOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encapsulates settings provided to GetIamPolicy.
}
\concept{GoogleIamV1GetPolicyOptions functions}
