% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleIamV1Policy}
\alias{GoogleIamV1Policy}
\title{GoogleIamV1Policy Object}
\usage{
GoogleIamV1Policy(
  auditConfigs = NULL,
  bindings = NULL,
  version = NULL,
  etag = NULL
)
}
\arguments{
\item{auditConfigs}{Specifies cloud audit logging configuration for this policy}

\item{bindings}{Associates a list of \code{members}, or principals, with a \code{role}}

\item{version}{Specifies the format of the policy}

\item{etag}{\code{etag} is used for optimistic concurrency control as a way to help prevent simultaneous updates of a policy from overwriting each other}
}
\value{
GoogleIamV1Policy object
}
\description{
GoogleIamV1Policy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A \code{Policy} is a collection of \code{bindings}. A \code{binding} binds one or more \code{members}, or principals, to a single \code{role}. Principals can be user accounts, service accounts, Google groups, and domains (such as G Suite). A \code{role} is a named list of permissions; each \code{role} can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a \code{binding} can also specify a \code{condition}, which is a logical expression that allows access to a resource only if the expression evaluates to \code{true}. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the \href{https://cloud.google.com/iam/help/conditions/resource-policies}{IAM documentation}. \strong{JSON example:} { 'bindings': [ { 'role': 'roles/resourcemanager.organizationAdmin', 'members': \link{ 'user:mike@example.com', 'group:admins@example.com', 'domain:google.com', 'serviceAccount:my-project-id@appspot.gserviceaccount.com' } }, { 'role': 'roles/resourcemanager.organizationViewer', 'members': \link{ 'user:eve@example.com' }, 'condition': { 'title': 'expirable access', 'description': 'Does not grant access after Sep 2020', 'expression': 'request.time < timestamp('2020-10-01T00:00:00.000Z')', } } ], 'etag': 'BwWWja0YfJA=', 'version': 3 } \strong{YAML example:} bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') etag: BwWWja0YfJA= version: 3 For a description of IAM and its features, see the \href{https://cloud.google.com/iam/docs/}{IAM documentation}.
}
\concept{GoogleIamV1Policy functions}
