% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleIamV1SetIamPolicyRequest}
\alias{GoogleIamV1SetIamPolicyRequest}
\title{GoogleIamV1SetIamPolicyRequest Object}
\usage{
GoogleIamV1SetIamPolicyRequest(updateMask = NULL, policy = NULL)
}
\arguments{
\item{updateMask}{OPTIONAL: A FieldMask specifying which fields of the policy to modify}

\item{policy}{REQUIRED: The complete policy to be applied to the \code{resource}}
}
\value{
GoogleIamV1SetIamPolicyRequest object
}
\description{
GoogleIamV1SetIamPolicyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for \code{SetIamPolicy} method.
}
\seealso{
Other GoogleIamV1SetIamPolicyRequest functions: 
\code{\link{projects.tenants.setIamPolicy}()}
}
\concept{GoogleIamV1SetIamPolicyRequest functions}
