% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleIamV1TestIamPermissionsRequest}
\alias{GoogleIamV1TestIamPermissionsRequest}
\title{GoogleIamV1TestIamPermissionsRequest Object}
\usage{
GoogleIamV1TestIamPermissionsRequest(permissions = NULL)
}
\arguments{
\item{permissions}{The set of permissions to check for the \code{resource}}
}
\value{
GoogleIamV1TestIamPermissionsRequest object
}
\description{
GoogleIamV1TestIamPermissionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for \code{TestIamPermissions} method.
}
\seealso{
Other GoogleIamV1TestIamPermissionsRequest functions: 
\code{\link{projects.tenants.testIamPermissions}()}
}
\concept{GoogleIamV1TestIamPermissionsRequest functions}
