% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GoogleProtobufEmpty}
\alias{GoogleProtobufEmpty}
\title{GoogleProtobufEmpty Object}
\usage{
GoogleProtobufEmpty()
}
\value{
GoogleProtobufEmpty object
}
\description{
GoogleProtobufEmpty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); }
}
\concept{GoogleProtobufEmpty functions}
