% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\docType{package}
\name{identitytoolkit_googleAuthR}
\alias{identitytoolkit_googleAuthR}
\title{Identity Toolkit API
The Google Identity Toolkit API lets you use open standards to verify a user's identity.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
at 2023-06-21 08:48:24.212216
filename: /Users/justinjm/projects/github.com/justinjm/autoGoogleAPI/googleidentitytoolkitv2.auto/R/identitytoolkit_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}
}
