% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.defaultSupportedIdpConfigs.create}
\alias{projects.defaultSupportedIdpConfigs.create}
\title{Create a default supported Idp configuration for an Identity Toolkit project.}
\usage{
projects.defaultSupportedIdpConfigs.create(
  GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig,
  parent,
  idpId = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig}{The \link{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig} object to pass to this method}

\item{parent}{The parent resource name where the config to be created, for example: 'projects/my-awesome-project'}

\item{idpId}{The id of the Idp to create a config for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig}()},
\code{\link{projects.defaultSupportedIdpConfigs.patch}()},
\code{\link{projects.tenants.defaultSupportedIdpConfigs.create}()},
\code{\link{projects.tenants.defaultSupportedIdpConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig functions}
