% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.defaultSupportedIdpConfigs.patch}
\alias{projects.defaultSupportedIdpConfigs.patch}
\title{Update a default supported Idp configuration for an Identity Toolkit project.}
\usage{
projects.defaultSupportedIdpConfigs.patch(
  GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig}{The \link{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig} object to pass to this method}

\item{name}{The name of the DefaultSupportedIdpConfig resource, for example: 'projects/my-awesome-project/defaultSupportedIdpConfigs/google}

\item{updateMask}{The update mask applies to the resource}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig}()},
\code{\link{projects.defaultSupportedIdpConfigs.create}()},
\code{\link{projects.tenants.defaultSupportedIdpConfigs.create}()},
\code{\link{projects.tenants.defaultSupportedIdpConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2DefaultSupportedIdpConfig functions}
