% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.identityPlatform.initializeAuth}
\alias{projects.identityPlatform.initializeAuth}
\title{Initialize Identity Platform for a Cloud project. Identity Platform is an end-to-end authentication system for third-party users to access your apps and services. These could include mobile/web apps, games, APIs and beyond. This is the publicly available variant of EnableIdentityPlatform that is only available to billing-enabled projects.}
\usage{
projects.identityPlatform.initializeAuth(
  GoogleCloudIdentitytoolkitAdminV2InitializeIdentityPlatformRequest,
  project
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2InitializeIdentityPlatformRequest}{The \link{GoogleCloudIdentitytoolkitAdminV2InitializeIdentityPlatformRequest} object to pass to this method}

\item{project}{The resource name of the target project the developer wants to enable Identity Platform for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2InitializeIdentityPlatformRequest functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2InitializeIdentityPlatformRequest}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2InitializeIdentityPlatformRequest functions}
