% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.inboundSamlConfigs.create}
\alias{projects.inboundSamlConfigs.create}
\title{Create an inbound SAML configuration for an Identity Toolkit project.}
\usage{
projects.inboundSamlConfigs.create(
  GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig,
  parent,
  inboundSamlConfigId = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig}{The \link{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig} object to pass to this method}

\item{parent}{The parent resource name where the config to be created, for example: 'projects/my-awesome-project'}

\item{inboundSamlConfigId}{The id to use for this config}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig}()},
\code{\link{projects.inboundSamlConfigs.patch}()},
\code{\link{projects.tenants.inboundSamlConfigs.create}()},
\code{\link{projects.tenants.inboundSamlConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2InboundSamlConfig functions}
