% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.oauthIdpConfigs.create}
\alias{projects.oauthIdpConfigs.create}
\title{Create an Oidc Idp configuration for an Identity Toolkit project.}
\usage{
projects.oauthIdpConfigs.create(
  GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig,
  parent,
  oauthIdpConfigId = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig}{The \link{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig} object to pass to this method}

\item{parent}{The parent resource name where the config to be created, for example: 'projects/my-awesome-project'}

\item{oauthIdpConfigId}{The id to use for this config}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig}()},
\code{\link{projects.oauthIdpConfigs.patch}()},
\code{\link{projects.tenants.oauthIdpConfigs.create}()},
\code{\link{projects.tenants.oauthIdpConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig functions}
