% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.oauthIdpConfigs.patch}
\alias{projects.oauthIdpConfigs.patch}
\title{Update an Oidc Idp configuration for an Identity Toolkit project.}
\usage{
projects.oauthIdpConfigs.patch(
  GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig}{The \link{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig} object to pass to this method}

\item{name}{The name of the OAuthIdpConfig resource, for example: 'projects/my-awesome-project/oauthIdpConfigs/oauth-config-id'}

\item{updateMask}{The update mask applies to the resource}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig}()},
\code{\link{projects.oauthIdpConfigs.create}()},
\code{\link{projects.tenants.oauthIdpConfigs.create}()},
\code{\link{projects.tenants.oauthIdpConfigs.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2OAuthIdpConfig functions}
