% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.tenants.create}
\alias{projects.tenants.create}
\title{Create a tenant. Requires write permission on the Agent project.}
\usage{
projects.tenants.create(GoogleCloudIdentitytoolkitAdminV2Tenant, parent)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2Tenant}{The \link{GoogleCloudIdentitytoolkitAdminV2Tenant} object to pass to this method}

\item{parent}{The parent resource name where the tenant will be created}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2Tenant functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers}()},
\code{\link{GoogleCloudIdentitytoolkitAdminV2Tenant}()},
\code{\link{projects.tenants.patch}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2Tenant functions}
