% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.tenants.delete}
\alias{projects.tenants.delete}
\title{Delete a tenant. Requires write permission on the Agent project.}
\usage{
projects.tenants.delete(name)
}
\arguments{
\item{name}{Resource name of the tenant to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}
}
