% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.tenants.getIamPolicy}
\alias{projects.tenants.getIamPolicy}
\title{Gets the access control policy for a resource. An error is returned if the resource does not exist. An empty policy is returned if the resource exists but does not have a policy set on it. Caller must have the right Google IAM permission on the resource.}
\usage{
projects.tenants.getIamPolicy(GoogleIamV1GetIamPolicyRequest, resource)
}
\arguments{
\item{GoogleIamV1GetIamPolicyRequest}{The \link{GoogleIamV1GetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy is being requested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleIamV1GetIamPolicyRequest functions: 
\code{\link{GoogleIamV1GetIamPolicyRequest}()}
}
\concept{GoogleIamV1GetIamPolicyRequest functions}
