% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.tenants.list}
\alias{projects.tenants.list}
\title{List tenants under the given agent project. Requires read permission on the Agent project.}
\usage{
projects.tenants.list(parent, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{The pagination token from the response of a previous request}

\item{pageSize}{The maximum number of results to return, capped at 1000}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}
}
