% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.tenants.patch}
\alias{projects.tenants.patch}
\title{Update a tenant. Requires write permission on the Tenant resource.}
\usage{
projects.tenants.patch(
  GoogleCloudIdentitytoolkitAdminV2Tenant,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2Tenant}{The \link{GoogleCloudIdentitytoolkitAdminV2Tenant} object to pass to this method}

\item{name}{Output only}

\item{updateMask}{If provided, only update fields set in the update mask}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2Tenant functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2Tenant.testPhoneNumbers}()},
\code{\link{GoogleCloudIdentitytoolkitAdminV2Tenant}()},
\code{\link{projects.tenants.create}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2Tenant functions}
