% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.tenants.setIamPolicy}
\alias{projects.tenants.setIamPolicy}
\title{Sets the access control policy for a resource. If the policy exists, it is replaced. Caller must have the right Google IAM permission on the resource.}
\usage{
projects.tenants.setIamPolicy(GoogleIamV1SetIamPolicyRequest, resource)
}
\arguments{
\item{GoogleIamV1SetIamPolicyRequest}{The \link{GoogleIamV1SetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy is being specified}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleIamV1SetIamPolicyRequest functions: 
\code{\link{GoogleIamV1SetIamPolicyRequest}()}
}
\concept{GoogleIamV1SetIamPolicyRequest functions}
