% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.tenants.testIamPermissions}
\alias{projects.tenants.testIamPermissions}
\title{Returns the caller's permissions on a resource. An error is returned if the resource does not exist. A caller is not required to have Google IAM permission to make this request.}
\usage{
projects.tenants.testIamPermissions(
  GoogleIamV1TestIamPermissionsRequest,
  resource
)
}
\arguments{
\item{GoogleIamV1TestIamPermissionsRequest}{The \link{GoogleIamV1TestIamPermissionsRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy detail is being requested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleIamV1TestIamPermissionsRequest functions: 
\code{\link{GoogleIamV1TestIamPermissionsRequest}()}
}
\concept{GoogleIamV1TestIamPermissionsRequest functions}
