% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{projects.updateConfig}
\alias{projects.updateConfig}
\title{Update an Identity Toolkit project configuration.}
\usage{
projects.updateConfig(
  GoogleCloudIdentitytoolkitAdminV2Config,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudIdentitytoolkitAdminV2Config}{The \link{GoogleCloudIdentitytoolkitAdminV2Config} object to pass to this method}

\item{name}{Output only}

\item{updateMask}{The update mask applies to the resource}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/identity-platform}{Google Documentation}

Other GoogleCloudIdentitytoolkitAdminV2Config functions: 
\code{\link{GoogleCloudIdentitytoolkitAdminV2Config}()}
}
\concept{GoogleCloudIdentitytoolkitAdminV2Config functions}
